Write a program to create a dictionary called accounts containing account id and #balance of account holders.
Initialize it with the following data:
101:50000     
102:45000           
103:55000
Now ask the user to input an account id and amount. If the account id is present
in the dictionary then update the balance by adding the amount given otherwise,
add a new entry of account id and balance in the dictionary.
Finally, print all the accounts details.